backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
qdrant-client
cohere
pydantic>=1.8,<2.0
pytest
openai>=1.87.0
nest_asyncio
asyncpg
SQLAlchemy
python-multipart
pypdf2
python-docx
tiktoken
beautifulsoup4
markdown